BugTracker - a new project
This project is an implementation of a previous tutorial (Basic Blog) into a standalone project. Previous work has been done (in the previous github project labeled "ArticleTracker") to have a good foundation for this project. Sentry.io integration, changing site design, and implementing a new app to django are a few of the changes to name a few